+++ lib/pf_putnbr_base.c
@@ -6,7 +6,7 @@
 /*   By: cmiran <cmiran@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/29 17:56:28 by cmiran            #+#    #+#             */
-/*   Updated: 2019/04/05 18:45:21 by cmiran           ###   ########.fr       */
+/*   Updated: 2019/04/07 17:21:54 by cmiran           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ unsigned char	itoc(size_t i, char b)
 	return (0);
 }
 
-void	pf_putnbr_base(size_t unbr, char b, size_t *fla, size_t *ret)
+void	pf_putnbr_base(unsigned long long unbr, char b, size_t *fla, size_t *ret)
 {
 	unsigned char	base;
